Fixed the type support issue

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -1,4 +1,5 @@
 #include <stdio.h>
+#include <stdint.h>
 
 /**
  * main - prints the first 50 fibonacci numbers
@@ -8,14 +9,14 @@
 int main(void)
 {
 	int limit = 50;
-	int num1 = 0;
-	int num2 = 1;
+	int64_t num1 = 0;
+	int64_t num2 = 1;
 	int i;
 
 	for (i = 0; i < limit; i++)
 	{
-		int temp = num1;
-	
+		int64_t temp = num1;
+
 		printf("%d%s", num1 + num2, i == limit - 1 ? "\n" : ", ");
 		num1 = num2;
 		num2 += temp;
